requirements.txt
flask==3.0.3
flask-cors==5.0.0
pyjwt==2.9.0